backend/requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Google Cloud libraries
google-cloud-logging==3.8.0
google-cloud-bigquery==3.13.0
google-cloud-bigquery-datatransfer==3.11.1
google-cloud-resource-manager==1.11.0
google-cloud-pubsub==2.18.4
google-cloud-storage==2.10.0
google-auth==2.23.4

# Utilities
python-dotenv==1.0.0
httpx==0.25.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.1
